Revert mitecom to python3
The mitecom has to work in python3 because some logic in
humanoid_league_team_communication is python3-only. Now, the behaviour
does not require the mitecom to use python2 anymore because the values
that were previously imported are now taken directly from the message
definition in humanoid_league_msgs/TeamData.