Move connector to bitbots_behaviour repository
The connector was moved to bitbots_behaviour because it is only ever
used in the head and body behaviour. Therefore, it makes sense to move
the package to this repository. The readability is also improved and it
is easier to commit changes because they are only in a single
repository.

Commits prior to this one can be found in the old location at
bitbots_misc/bitbots_common/connector.