ros_control: add delay before motor write after power is turned on
This is necessary because the motors need some time for their startup.
When ROM/RAM values are written directly after startup, they don't take
effect. The start_delay parameter is added to set this delay.
From testing, a value of 10ms would be sufficient but I chose a higher
value to be on the safe side and because this is only rarely executed.