feat(behavior): descend with goalie on block/role-position
by creating a generalized cancelable `AbstractDynupAnimation` class,
where each implementing class has to define a `reset_animation` method,
which is triggered on pop.

`GetWalkready` is also now using this as a base in our behavior.

In the future this class should be merge with the similar
`PlayAnimationDynup` class in HCM.

We now descend as goalie both on our role position and when we are in
the block position, which is handled by the `OnBlockPosition` decision.